[log-middleware] Add remote addr and port.
Context
-------

Currently, Traefik doesn't log the remote IP (or client
IP). However, some use cases involve extracting the client IP from
request logs:
  * use of GeoIP to build a world map dashboard showing
    where requests are coming from.
  * security audit logs / DDoS attack analysis
  * ...

This commit prints the remote IP address in the logs, so that they can
later be parsed.

Note to the reviewer/merger: this is likely to be compatibility
breaking for people that are already parsing Traefik's logs.